pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scipilot"
version = "0.1.0"
description = "SciPilot - Natural language interface for scientific CLI tools via MCP"
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
]
license = {file = "LICENSE"}
authors = [
    {name = "Dimitry Grebenyuk", email = "dimitrygrebenyuk@icloud.com"}
]
keywords = ["mcp", "scientific", "cli", "automation", "tools"]
dependencies = [
    "fastmcp>=2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "types-PyYAML>=6.0.12.20240311",
]

[project.scripts]
scipilot = "scipilot.server:main"

[project.urls]
Homepage = "https://github.com/grebenyyk/scipilot"
Repository = "https://github.com/grebenyyk/scipilot"
Issues = "https://github.com/grebenyyk/scipilot/issues"
Security = "https://github.com/grebenyyk/scipilot/security"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.13"
strict = true

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true
